requirements.txt
aniso8601==8.0.0
# appdirs==1.4.4
# autopep8==1.5.4
# Bootstrap-Flask==1.5
click==7.1.2
# distlib==0.3.1
# filelock==3.0.12
# flake8==3.8.3
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
# Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
# mccabe==0.6.1
# mypy==0.782
# mypy-extensions==0.4.3
# pycodestyle==2.6.0
# pyflakes==2.2.0
# pyreadline==2.1
pytz==2020.1
six==1.15.0
# SQLAlchemy==1.3.19
# toml==0.10.1
# typed-ast==1.4.1
# typing-extensions==3.7.4.3
# virtualenv==20.0.31
Werkzeug==1.0.1
# WTForms==2.3.3